make serialize_binary_path forgiving and move method location
